move render_message_history to message_handler.rs
